pyproject.toml
[project]
name = "crossvector"
version = "0.1.0"
description = "Cross-platform vector database engine with pluggable adapters"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Two Dev",email = "thetwofarm@gmail.com"}
]
maintainers = [
  { name = "Two Dev", email = "thetwofarm@gmail.com" },
]
keywords = ["vector-database", "embeddings", "semantic-search", "astradb", "chromadb", "milvus", "pgvector", "cross-platform", "llm", "rag"]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Development Status :: 4 - Beta",
]

dependencies = [
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Vector database adapters
astradb = ["astrapy>=2.1.0"]
chromadb = ["chromadb>=1.3.4"]
milvus = ["pymilvus>=2.6.3"]
pgvector = ["pgvector>=0.4.1", "psycopg2-binary>=2.9.11"]

# Embedding providers
openai = ["openai>=2.6.1"]
gemini = ["google-genai>=0.3.0"]

# Install all adapters
all-dbs = [
    "astrapy>=2.1.0",
    "chromadb>=1.3.4",
    "pymilvus>=2.6.3",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
]

# Install all embedding providers
all-embeddings = [
    "openai>=2.6.1",
    "google-genai>=0.3.0",
]

# Install everything
all = [
    "astrapy>=2.1.0",
    "chromadb>=1.3.4",
    "pymilvus>=2.6.3",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "openai>=2.6.1",
    "google-genai>=0.3.0",
]

# Development dependencies
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.18.2",
    "ruff>=0.13.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pre-commit>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/thewebscraping/crossvector"
Repository = "https://github.com/thewebscraping/crossvector"
Documentation = "https://github.com/thewebscraping/crossvector#readme"
"Bug Tracker" = "https://github.com/thewebscraping/crossvector/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crossvector"]

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py"]
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=crossvector --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
allow_untyped_calls = true
allow_untyped_defs = true
no_implicit_optional = true
